Made some changes in code

+++ encrypt.py
@@ -20,6 +20,23 @@ def encrypt_folder(folder_path):
                 data = f.read()
 
             encrypted = fernet.encrypt(data)
+    for file in os.listdir(folder_path):
+        file_path = os.path.join(folder_path, file)
+
+        if os.path.isfile(file_path):
+            with open(file_path, "rb") as f:
+                data = f.read()
+
+            encrypted = fernet.encrypt(data)
+for file in os.listdir(folder_path):
+        file_path = os.path.join(folder_path, file)
+
+        if os.path.isfile(file_path):
+            with open(file_path, "rb") as f:
+                data = f.read()
+
+            encrypted = fernet.encrypt(data)
+
 
             with open(file_path, "wb") as f:
                 f.write(encrypted)
